Pr feature

+++ app/Http/Controllers/ArrivalController.php
@@ -107,12 +107,32 @@ class ArrivalController extends Controller
     public function destroy(int $id): JsonResponse
     {
         $arrival = Arrival::findOrFail($id);
+
+        $kode = $arrival->kode;
+
+        $count = Arrival::where('kode', $kode)->count();
+
+        if ($count <= 1) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Gagal hapus, pastikan data tidak terintegrasi dengan data lainnya.'
+            ], 409);
+        }
+
         $arrival->delete();
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Data kedatangan berhasil dihapus.',
+            'message' => 'Data kedatangan berhasil dihapus.'
         ]);
+
+        // $arrival = Arrival::findOrFail($id);
+        // $arrival->delete();
+
+        // return response()->json([
+        //     'status' => 'success',
+        //     'message' => 'Data kedatangan berhasil dihapus.',
+        // ]);
     }
 
     public function deleteMultiple(Request $request): JsonResponse
